:art: (⚛️ Frontend) component file rename to match convention
