+++ client/src/components/Receipt/ItemsList/ItemsList.css
