+++ client/src/components/Receipt/PriceAndQR/priceAndQR.css
