+++ client/src/components/Receipt/ReceiptBody/ReceiptBody.css
@@ -88,7 +88,7 @@
 	font-size: small;
 }
 
-.List-Button {
+.listButton {
 	display: flex;
 	justify-content: center;
 }
