--- client/src/containers/receipt/header/header.css
@@ -1,10 +1,0 @@
-.Header {
-}
-
-.Header-Title {
-	text-align: center;
-	font-weight: 500;
-	margin-top: 2.5em;
-	margin-bottom: 2.5em;
-	color: white;
-}
